requirements.txt
streamlit
google-genai
pydub
